+++ src/main/java/com/hbm/entity/projectile/EntityCog.java
@@ -143,8 +143,8 @@ public class EntityCog extends EntityThrowableInterp {
     }
 
     @Override
-    public double getGravityVelocity() {
-        return inGround ? 0 : 0.03F;
+    public float getGravityVelocity() {
+        return inGround ? 0F : 0.03F;
     }
 
     @Override
